WP7: Use Background Transfer to download podcast episodes.
This is the correct way to download large files on the WP. This fixes the issue of writing the data to the disk after it's being downloaded.

We still need to check and probably add stuff to take care of all the special cases that can happen when starting a transfer.

Also, this commit does not include resumes from thumbstone.