WP7: Handling error when starting new background transfer.
For some reason the application download request limit is reached, although whenever the request completes, we remove the transfer request.

So we handle the exception we get when we add a new request by then removing all download requests that the system knows about us. We have always only one download request active, so when we initiate a new one, there shouldn't be others active. If, however, the system says we reached some limit, we just remove all requests from our app.