Add more error checking and Toast showing.
This also adds a reference to the XNA framework in hopes of fixing the audio playback for the version installed from Marketplace. Let's see. If this is the reason for the crashes, I say Microsoft should do something about this - it must be a bug. And so frustrating when all works on the local device and on the SW that's deployed from Visual Studio. How could a developer find out about this issue if all works on his/her development machine, but then suddenly some capability is not detected by the marketplace and it just starts to crash. And the crashes are on application versions that do not ouput any debug logs so tough luck for me who try to fix it. Anyway, let's hope this commit finally fixes the crash and we can move along.