Fix crash in WP8?
When we enter the player via back button, it might be that the screen updates are coming before we initialize the player. Hence there is a null pointer because the player instance doesn't have a track yet.

This commit tries to fix that by removing the timer when we don't show the player, and starts it only when it becomes visible.