Fix crash: Showing error toast from ongoing background transfer caused crash.
This is because both the download manager and the toast showing code was in the static part of App class. So when the download manager tries to show an error toast when it processed the download request which hasn't yet been completed, it crashes because the showToast() method had not yet been initialized.