Fix issue of lost Downloading state.
If the download is started when we are on cellualr connection, but state then goes to "Waiting for Wifi...", the previous order in this case made it so that the transfer request is never set back for this episode. This is because the NotifyPropertyChanged("ProgressBarValue") calls ProgressBar.Get which is responsible for hooking up the download progress values from the transfer request. But it checked if the transfer request was null before setting it up - which it wasn't earlier.