First version of play queue functionality.
One can now actually see the play queue as sent to the audio player, as well as tap on an item to start the playback of another item.

Unfortunately this code is so ugly that it makes your eyes bleed. We heavily need to refactor how podcast playback is started; at least unify how it's started from the episode listing and here. Now the case is that we need to update the play queue data into the DB from two places, which is just wrong, and I don't like how we have to initiate the player control explicitely. A big TODO.